megano/requirements.txt
asgiref==3.5.2
attrs==23.2.0
black==24.2.0
build==1.1.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
diploma-frontend @ file:///Users/dorinavancea/PycharmProjects/python_django_diploma2/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=0e5e5ecdfc8c35206c8065257940d7428f072264d47a1a54fc5e86bc9f643f6a
distlib==0.3.7
Django==4.1.6
django-debug-toolbar==4.3.0
django-filter==22.1
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.27.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
importlib_metadata==7.0.2
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.4.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pytz==2022.7.1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.17.1
sentry-sdk==1.40.5
shellingham==1.5.4
sqlparse==0.4.2
tomlkit==0.12.4
trove-classifiers==2024.3.3
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.24.6
xattr==1.1.0
zipp==3.18.1
